move timeline, fix intervals in progress bar
